Fix iteration over out parameters after invoking a function.
We were using the same index to iterate over both in and out arrays,
but since they are different arrays both starting at 0 we need to use
two variables.

On top of that, use out_args as the arguments source, since that's the
array that we pass to g_function_info_invoke.