[libseed] Add more attractive way to define GObject properties on newly created GTypes.
This is incomplete as of yet (only supports bool and char properties, but it's mostly just copy-'n-paste from here on out).

Also, add a test that shows how to use this (to add a single boolean property, the others follow in an obvious way, except add "maximum_value" and "minimum_value" properties for properties that use them.

This also needs a *ton* more error checking.