libseed: Properly bubble exceptions raised during class init
Before, we were simply catching the exception, printing it, and continuing
on happily. This completely defeats the point of exceptions. Now, we pop
the exception into the class's qdata (a little ugly) and bubble it up and
clear it the first time we encounter it (when the class is created, the
first time it's instantiated).