Fix subtle crashed when a dbus reply has no message content.
In the DBus spec, its perfectly valid to send empty reply messages, even
if a method call message has been set 'no_reply'. In this case, the dbus
module was invoking a closure with uninitialised data, causing all sorts
of interesting behaviour.

This patch fixes this by not invoking the
closure when a reply is empty.