Allow the creation of a restricted/constrained SeedEngine and add APIs to expose choosen GObject namspaces and instance into this engine
The Javascript context of a restricted SeedEngine is empty by default.
The importer module, builtins and GType overriding facilities aren't
exposed to JS Code.
The developer must expose *explicitely* the objects and namespaces
that should be accessible from the Javascript context.

Add hight-level API to expose GObject instance and namespace in the JS context

https://bugzilla.gnome.org/show_bug.cgi?id=645696