Adding ondemand for reference -
Standard seed loads all properties into the namespace when you reference it, this patch file changes the behaviour so a fake namespace object
is created, then when you request properties, the objects are created (eg. ctors/classes/enums etc..)

The net effect of this is a shared memory reduction of around 20M when using a few key namespaces (eg. Gtk ... etc..)

Still not close to gjs memory but a small improvement..