closure: Don't call JSValueUnprotect() from a finalizer
We would not get here at all anyway if the object still had a "protect count"
higher than 0, as that would prevent the GC from cleaning up the object.

Also the docs say:

You must not call any function that may cause a garbage collection or an allocation
of a garbage collected object from within a JSObjectFinalizeCallback. This includes
all functions that have a JSContextRef parameter.

https://bugzilla.gnome.org/show_bug.cgi?id=742574