Made GLib variable in seed.js scoped
In seed.js the GLib variable is global. This is no problem as long
there is no const or let keyword in jsc.
After adding this keywords by supporting the new webkit2gtk4.0 port,
you get the following error.

SyntaxError Can't create duplicate variable that shadows a global
property: 'GLib'

This patch make the GLib variable const which also limit the scope
of it.

https://bugzilla.gnome.org/show_bug.cgi?id=759171