Add support for the global ARGV object
In seed command line parameter will be provided by the Seed.argv
object. In gjs this is done with the global ARGV object. Also,
the count of parameter is different. While in seed the command
line is completly provided, in gjs the first 2 parameter are obmitted.

This patch add support for the ARGV object in seed.

https://bugzilla.gnome.org/show_bug.cgi?id=759171