Add gjs compatibility layer to seed
This Adds the following:
 * Add all js modules and extensions from gjs to the
   seed codebase.
 * Add a compile time flag to build with and without support.
 * Add a runtime commandline to switch compatibility mode on and off.
 * Loads the gjs modules first if commandline switch is given and do
   not load seed modules if a gjs module is present.

Note:
Not all gjs modules are working because of missing native modules in seed.
There are: window, system, cairo and console.
They need to be ported.
The gjs modules are expected to start working when they are ported.

This patch will not change current behaviour of seed when
--seed-gjs-compatibility commandline is not given or when gjs support
is not build in.
Current seed code should still work.