Add the native gjs module dummy "system" for compatibility
This is just a dummy. The functions should be implemented.
It shows how to implement native gjs compatibility modules.
It has the flaw, that the js compatibility modules do not recognice
native gjs compatibility  modules, this should be solved.
I get this running gjs code:

** (seed:27750): CRITICAL **: Line 49 in
/usr/local/share/seed-gtk4/gjs/package.js:
TypeError undefined is not an object
(evaluating 'System.programInvocationName')

It do not harm to integrate this
patch, even when it is just a dummy.