Moving all JS files to a GResource to improve uninstalled setup.
At the current moment, uninstalled usage is really hard with seed.
If you try to run with: ./src/seed it runs, but it tries to read the
javascript files and native modules from the INSTALL PATH.

This commit solve that issue for the javascript files, however native
modules still have the same issue.