doc/modules: gtk-doc must be run in the html subfolder.
gtkdoc-mkhtml will output html.stamp in the parent folder, and
it never got cleaned afterwards.

Not sure why doc/modules does not use the gtk-doc makefile
fragment, but I'm not going for the best solution, simply the
one that does not break distcheck by leaving build artefacts in
random places.

https://bugzilla.gnome.org/show_bug.cgi?id=771510